Created a new Component to house the Users Panel
Replaced the current Users Panel with a self-closing component tag.

Wrapped the component tag in RequireGroup

removed now unncessary states/constants and imports from Users.tsx
